requirements.txt
flask
deta==0.9
python-dotenv
werkzeug
pytz
flask-cors
Flask-WeasyPrint